frittståendeBrevService skal ha ansvar for generering av automatiske brev for innhenting av karakterutskrift
